+++ spec/controllers/elements_controller_spec.rb
@@ -24,25 +24,25 @@ RSpec.describe ElementsController, type: :controller do
     datatable = ElementsDatatable.new
     
     describe "filters:" do
-      datatable.filter_keys.each do |key|
+      datatable.searchable_columns.each do |key|
         next if key == :goe  # TODO
         it key do
           expect_filter(elem4T, elemLSp, key)
         end
       end
       context 'element name with perfect match' do
-        subject { get :list, xhr: true, params: {name: '4T', perfect_match: 'PERFECT_MATCH'} }
+        subject { get :list, xhr: true, params: {name: '4T', name_operator: 'eq'} }
         its(:body) { is_expected.to include(elem4T.name) }
         its(:body) { is_expected.not_to include(elem4T3T.name) }
       end
       context 'comparison goe by' do
         context '>' do
-          subject { get :list, xhr: true, params: {goe: '1', goe_operator: '>'} }
+          subject { get :list, xhr: true, params: {goe: '1', goe_operator: 'gt'} }
           its(:body) { is_expected.to include(elem4T.name) }
           its(:body) { is_expected.not_to include(elemLSp.name) }
         end
         context '<' do
-          subject { get :list, xhr: true, params: {goe: '1', goe_operator: '<'} }
+          subject { get :list, xhr: true, params: {goe: '1', goe_operator: 'lt'} }
           its(:body) { is_expected.not_to include(elem4T.name) }
           its(:body) { is_expected.to include(elemLSp.name) }
         end